+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/CollectorRestController.java
@@ -2,11 +2,12 @@ package com.nimio2024.nimio2024_tfg_polmorenocastelar.application;
 
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.exceptions.StudentNotExistsException;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.controllers.CollectorController;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Collector;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Person;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Student;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/collectors")
@@ -21,6 +22,11 @@ public class CollectorRestController {
     @PutMapping("/{studentId}/auth/{personId}")
     public Student addPersonToStudent(@PathVariable Long studentId, @PathVariable Long personId) throws StudentNotExistsException {
         return collectorController.addPersonToStudent(studentId, personId);
+    }
+
+    @GetMapping("/auth/{studentId}")
+    public List<Collector> getCollectorsPersons(@PathVariable Long studentId) {
+        return collectorController.getCollectorsPersonsByStudentId(studentId);
     }