+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/StudentRestController.java
@@ -61,6 +61,7 @@ public class StudentRestController {
         studentController.deleteAuth(personId, studentId);
     }
 
+
     @DeleteMapping("/{studentId}")
     public void deleteStudent(@PathVariable Long studentId) {
         studentController.deleteStudent(studentId);